Move TabGroupData and associated enum to separate file
