+++ Client/Frontend/Browser/SavedTab.swift
@@ -90,54 +90,3 @@ class SavedTab: NSObject, NSCoding {
         coder.encode(hasHomeScreenshot, forKey: "hasHomeScreenshot")
     }
 }
-
-public class TabGroupData: NSObject, NSCoding {
-    var tabAssociatedSearchTerm: String = ""
-    var tabAssociatedSearchUrl: String = ""
-    var tabAssociatedNextUrl: String = ""
-    var tabHistoryCurrentState = ""
-    var tabGroupTimerState = ""
-    
-    // Checks if there is any search term data
-    func isEmpty() -> Bool {
-        return tabAssociatedSearchTerm.isEmpty && tabAssociatedSearchUrl.isEmpty && tabAssociatedNextUrl.isEmpty
-    }
-
-   func tabHistoryMetadatakey() -> HistoryMetadataKey {
-       return HistoryMetadataKey(url: tabAssociatedSearchUrl, searchTerm: tabAssociatedSearchTerm, referrerUrl: tabAssociatedNextUrl)
-   }
-    
-    var jsonDictionary: [String: Any] {
-        return [
-            "tabAssociatedSearchTerm": String(self.tabAssociatedSearchTerm),
-            "tabAssociatedSearchUrl": String(self.tabAssociatedSearchUrl),
-            "tabAssociatedNextUrl": String(self.tabAssociatedNextUrl),
-            "tabHistoryCurrentState": String(self.tabHistoryCurrentState),
-            "tabGroupTimerState": String(self.tabGroupTimerState),
-        ]
-    }
-
-    init(searchTerm: String, searchUrl: String, nextReferralUrl: String, tabHistoryCurrentState: String, tabGroupTimerState: String) {
-        self.tabAssociatedSearchTerm = searchTerm
-        self.tabAssociatedSearchUrl = searchUrl
-        self.tabAssociatedNextUrl = nextReferralUrl
-        self.tabHistoryCurrentState = tabHistoryCurrentState
-        self.tabGroupTimerState = tabGroupTimerState
-    }
-
-    required public init?(coder: NSCoder) {
-        self.tabAssociatedSearchTerm = coder.decodeObject(forKey: "tabAssociatedSearchTerm") as? String ?? ""
-        self.tabAssociatedSearchUrl = coder.decodeObject(forKey: "tabAssociatedSearchUrl") as? String ?? ""
-        self.tabAssociatedNextUrl = coder.decodeObject(forKey: "tabAssociatedNextUrl") as? String ?? ""
-        self.tabHistoryCurrentState = coder.decodeObject(forKey: "tabHistoryCurrentState") as? String ?? ""
-        self.tabGroupTimerState = coder.decodeObject(forKey: "tabGroupTimerState") as? String ?? ""
-    }
-
-    public func encode(with coder: NSCoder) {
-        coder.encode(tabAssociatedSearchTerm, forKey: "tabAssociatedSearchTerm")
-        coder.encode(tabAssociatedSearchUrl, forKey: "tabAssociatedSearchUrl")
-        coder.encode(tabAssociatedNextUrl, forKey: "tabAssociatedNextUrl")
-        coder.encode(tabHistoryCurrentState, forKey: "tabHistoryCurrentState")
-        coder.encode(tabGroupTimerState, forKey: "tabGroupTimerState")
-    }
-}
